+++ example/lib/general_page/general_page.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:better_player/better_player.dart';
+import 'package:better_player_example/other_page/other_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:path_provider/path_provider.dart';
@@ -18,17 +19,22 @@ class _GeneralPageState extends State<GeneralPage> {
   bool _fileVideoShown = false;
 
   Future<BetterPlayerController> _setupDefaultVideoData() async {
-    var dataSource = BetterPlayerDataSource(BetterPlayerDataSourceType.NETWORK,
-        "https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ForBiggerBlazes.mp4",
-        subtitles: BetterPlayerSubtitlesSource(
-            type: BetterPlayerSubtitlesSourceType.NETWORK,
-            url:
-                "https://dl.dropboxusercontent.com/s/71nzjo2ux3evxqk/example_subtitles.srt"));
+    var dataSource = BetterPlayerDataSource(
+      BetterPlayerDataSourceType.NETWORK,
+      "https://rtmp.api.rt.com/hls/rtdru.m3u8",
+      liveStream: true,
+      subtitles: BetterPlayerSubtitlesSource(
+        type: BetterPlayerSubtitlesSourceType.NETWORK,
+        url:
+            "https://dl.dropboxusercontent.com/s/71nzjo2ux3evxqk/example_subtitles.srt",
+      ),
+    );
     _betterPlayerController = BetterPlayerController(
         BetterPlayerConfiguration(
-          controlsConfiguration:
-              BetterPlayerControlsConfiguration(enableProgressText: true),
-        ),
+            controlsConfiguration: BetterPlayerControlsConfiguration(
+          enableProgressText: true,
+          enablePlaybackSpeed: true,
+        )),
         betterPlayerDataSource: dataSource);
     _betterPlayerController.addEventsListener((event) {
       print("Better player event: ${event.betterPlayerEventType}");
@@ -64,10 +70,10 @@ class _GeneralPageState extends State<GeneralPage> {
   }
 
   Future _saveAssetVideoToFile() async {
-    String content = await rootBundle.loadString("assets/testvideo.mp4");
+    var content = await rootBundle.load("assets/testvideo.mp4");
     final directory = await getApplicationDocumentsDirectory();
-    var file = File("${directory.path}/textvideo.mp4");
-    file.writeAsString(content);
+    var file = File("${directory.path}/testvideo.mp4");
+    file.writeAsBytesSync(content.buffer.asUint8List());
   }
 
   @override
@@ -79,7 +85,8 @@ class _GeneralPageState extends State<GeneralPage> {
             " URL. Subtitles are loaded from file."),
       ),
       _buildDefaultVideo(),
-      _buildShowFileVideoButton()
+      _buildShowFileVideoButton(),
+      _buildOtherPageButton(),
     ]);
   }
 
@@ -88,7 +95,9 @@ class _GeneralPageState extends State<GeneralPage> {
       future: _setupDefaultVideoData(),
       builder: (context, snapshot) {
         if (!snapshot.hasData) {
-          return Center(child: CircularProgressIndicator());
+          return Center(
+            child: CircularProgressIndicator(),
+          );
         } else {
           return AspectRatio(
             aspectRatio: 16 / 9,
@@ -108,6 +117,21 @@ class _GeneralPageState extends State<GeneralPage> {
         onPressed: () {
           _fileVideoShown = !_fileVideoShown;
           _fileVideoStreamController.add(_fileVideoShown);
+        },
+      ),
+      _buildFileVideo()
+    ]);
+  }
+
+  Widget _buildOtherPageButton() {
+    return Column(children: [
+      RaisedButton(
+        child: Text("Show video in other page"),
+        onPressed: () {
+          Navigator.push(
+            context,
+            MaterialPageRoute(builder: (context) => OtherPage()),
+          );
         },
       ),
       _buildFileVideo()