+++ lib/src/configuration/better_player_event_type.dart
@@ -9,4 +9,7 @@ enum BetterPlayerEventType {
   PROGRESS,
   FINISHED,
   EXCEPTION,
+  CONTROLS_VISIBLE,
+  CONTROLS_HIDDEN,
+  SET_SPEED,
 }