+++ lib/src/controls/better_player_controls_state.dart
@@ -1,0 +1,136 @@
+import 'package:better_player/better_player.dart';
+import 'package:better_player/src/controls/better_player_clickable_widget.dart';
+import 'package:better_player/src/video_player/video_player.dart';
+import 'package:flutter/material.dart';
+
+///Base class for both material and cupertino controls
+abstract class BetterPlayerControlsState<T extends StatefulWidget>
+    extends State<T> {
+  ///Min. time of buffered video to hide loading timer (in milliseconds)
+  static const int _bufferingInterval = 20000;
+
+  BetterPlayerController getBetterPlayerController();
+
+  void onShowMoreClicked() {
+    showModalBottomSheet(
+      context: context,
+      builder: (context) {
+        return SafeArea(
+          top: false,
+          bottom: true,
+          child: _buildMoreOptionsList(),
+        );
+      },
+    );
+  }
+
+  Widget _buildMoreOptionsList() {
+    return SingleChildScrollView(
+      child: Container(
+        child: Column(
+          children: [
+            _buildMoreOptionsListRow(Icons.shutter_speed, "Playback speed", () {
+              Navigator.of(context).pop();
+              _showSpeedChooserWidget();
+            })
+          ],
+        ),
+      ),
+    );
+  }
+
+  Widget _buildMoreOptionsListRow(IconData icon, String name, Function onTap) {
+    assert(icon != null, "Icon can't be null");
+    assert(name != null, "Name can't be null");
+    assert(onTap != null, "OnTap can't be null");
+    return BetterPlayerMaterialClickableWidget(
+      onTap: onTap,
+      child: Padding(
+        padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 8),
+        child: Row(
+          children: [
+            Icon(icon),
+            const SizedBox(width: 16),
+            Text(name),
+          ],
+        ),
+      ),
+    );
+  }
+
+  void _showSpeedChooserWidget() {
+    showModalBottomSheet(
+        context: context,
+        builder: (context) {
+          return SafeArea(
+            top: false,
+            bottom: true,
+            child: SingleChildScrollView(
+              child: Column(
+                children: [
+                  _buildSpeedRow(0.25),
+                  _buildSpeedRow(0.5),
+                  _buildSpeedRow(0.75),
+                  _buildSpeedRow(1.0),
+                  _buildSpeedRow(1.25),
+                  _buildSpeedRow(1.5),
+                  _buildSpeedRow(1.75),
+                  _buildSpeedRow(2.0),
+                ],
+              ),
+            ),
+          );
+        });
+  }
+
+  Widget _buildSpeedRow(double value) {
+    assert(value != null, "Value can't be null");
+    return BetterPlayerMaterialClickableWidget(
+      child: Padding(
+        padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 8),
+        child: Row(
+          children: [
+            const SizedBox(width: 16),
+            Text(
+              "$value x",
+              style: TextStyle(
+                  fontWeight: getBetterPlayerController()
+                              .videoPlayerController
+                              .value
+                              .speed ==
+                          value
+                      ? FontWeight.bold
+                      : FontWeight.normal),
+            )
+          ],
+        ),
+      ),
+      onTap: () {
+        Navigator.of(context).pop();
+        getBetterPlayerController().setSpeed(value);
+      },
+    );
+  }
+
+  bool isLoading(VideoPlayerValue latestValue) {
+    assert(latestValue != null, "Latest value can't be null");
+    if (latestValue != null) {
+      if (!latestValue.isPlaying && latestValue.duration == null) {
+        return true;
+      }
+
+      var position = latestValue.position;
+      var bufferedEndPosition = latestValue.buffered.last.end;
+      if (position != null && bufferedEndPosition != null) {
+        var difference = bufferedEndPosition - position;
+
+        if (latestValue.isPlaying &&
+            latestValue.isBuffering &&
+            difference.inMilliseconds < _bufferingInterval) {
+          return true;
+        }
+      }
+    }
+    return false;
+  }
+}