+++ lib/src/list/better_player_list_video_player_controller.dart
@@ -16,6 +16,10 @@ class BetterPlayerListVideoPlayerController {
     _betterPlayerController?.play();
   }
 
+  void seekTo(Duration duration) {
+    _betterPlayerController?.seekTo(duration);
+  }
+
   void setBetterPlayerController(
       BetterPlayerController betterPlayerController) {
     _betterPlayerController = betterPlayerController;